+++ packages/prop-house-webapp/src/components/RoundCard/index.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from 'react-router-dom';
 import AwardLabels from '../AwardLabels';
 import { useState } from 'react';
 import Modal from '../Modal';
-import useFullRoundAwards from '../../hooks/useFullRoundAward';
+import useFullRoundAwards from '../../hooks/useFullRoundAwards';
 import LoadingIndicator from '../LoadingIndicator';
 import RoundCardStatusBar from '../RoundCardStatusBar';
